+++ app/services/activities/import_from_csv.rb
@@ -167,6 +167,7 @@ module Activities
         actual_start_date: "Actual start date",
         actual_end_date: "Actual end date",
         sector: "Sector",
+        channel_of_delivery_code: "Channel of delivery code",
         collaboration_type: "Collaboration type (Bi/Multi Marker)",
         flow: "Flow",
         aid_type: "Aid type",
@@ -304,6 +305,16 @@ module Activities
         )
       end
 
+      def convert_channel_of_delivery_code(channel_of_delivery_code)
+        raise I18n.t("importer.errors.activity.invalid_channel_of_delivery_code") if channel_of_delivery_code.blank?
+
+        validate_channel_of_delivery_code(
+          channel_of_delivery_code,
+          :channel_of_delivery_code,
+          I18n.t("importer.errors.activity.invalid_channel_of_delivery_code"),
+        )
+      end
+
       def convert_collaboration_type(collaboration_type)
         validate_from_codelist(
           collaboration_type,
@@ -390,6 +401,18 @@ module Activities
 
         codelist = load_yaml(entity: :activity, type: entity)
         valid_codes = codelist.map { |entry| entry.fetch("code") }
+
+        raise message unless valid_codes.include?(code)
+
+        code
+      end
+
+      def validate_channel_of_delivery_code(code, entity, message)
+        return nil if code.blank?
+
+        codelist = load_yaml(entity: :activity, type: entity)
+        valid_codes = codelist.map { |entry| entry.fetch("code") }
+        valid_codes << "N/A"
 
         raise message unless valid_codes.include?(code)