+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_11_25_172731) do
+ActiveRecord::Schema.define(version: 2020_11_30_114052) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -82,6 +82,7 @@ ActiveRecord::Schema.define(version: 2020_11_25_172731) do
     t.string "oda_eligibility_lead"
     t.string "beis_id"
     t.string "uk_dp_named_contact"
+    t.string "channel_of_delivery_code"
     t.index ["extending_organisation_id"], name: "index_activities_on_extending_organisation_id"
     t.index ["level"], name: "index_activities_on_level"
     t.index ["organisation_id"], name: "index_activities_on_organisation_id"